requirements.txt
yfinance
google-adk